Move Tasks to larq_zoo.experiments
This is in order to import larq zoo without poluting experiments of another zookeeper project.
